+++ CHANGELOG.md
@@ -14,8 +14,6 @@ For information on changes in released versions of Teku, see the [releases page]
 ### Breaking Changes
 
 ### Additions and Improvements
-* Updated CLI options ensuring unmatched options aren't confused as parameters.
+* Added `kintsugi` network definition. 
 
 ### Bug Fixes
-* Updated to log4j 2.16.0.
-* Fix multiarch JDK17 variant docker image to bundle Java 17 instead of Java 16